Change around routers a little bit and fix map
