+++ app.js
@@ -7,12 +7,17 @@ var logger = require('morgan');
 var sassMiddleware = require('node-sass-middleware');
 
 var indexRouter = require('./routes/index');
-var usersRouter = require('./routes/users');
+var ticketsRouter = require('./routes/tickets');
+
 
 var app = express();
 
 /* Routing to specific JS file */
-var tickets = require('./routes/tickets')(app);
+var ticketsRouter = require('./routes/tickets');
+var mapRouter = require('./routes/map');
+
+ticketsRouter(app);
+mapRouter(app);
 
 // view engine setup
 app.set('views', path.join(__dirname, 'views'));
@@ -32,7 +37,6 @@ app.use(sassMiddleware({
 app.use(express.static(path.join(__dirname, 'public')));
 
 app.use('/', indexRouter);
-app.use('/users', usersRouter);
 
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
