+++ views/map.hbs
@@ -3,6 +3,7 @@
     <link rel="stylesheet" href="/stylesheets/header.css">
     <link rel="stylesheet" href="/stylesheets/map.css">
     <link rel="stylesheet" href="/stylesheets/footer.css">
+    <script src="/js/map.js"></script>
 </head>
 <body>
 {{> header}}
